Shrink Entry count to u32 so it fits a single cache line
Entry was 40 bytes (count: u64), so an 8-aligned slot frequently
straddled a 64-byte cache line and a probe could fault two lines.
A per-station count is bounded by the 1e9-row total, well under
u32::MAX (~4.29e9), so count: u32 is safe and drops Entry to 32 B
(measured) — slots now pack two per line and never straddle.

No effect on the canonical 413-station file (the table stays in
L1/L2 either way), but ~10% faster on a 10,000-station file where
the table spills out of L1.

+++ src/main.rs
@@ -47,7 +47,11 @@ struct Entry {
     hash: u64,
     min: i16,
     max: i16,
-    count: u64,
+    // Per-station row count. u32 holds up to ~4.29e9, comfortably above the
+    // 1e9-row total even if one station dominated every line. Keeping the
+    // struct at 32 B (not 40) means an 8-aligned `Entry` never straddles a
+    // 64-byte cache line, so each probe touches a single line.
+    count: u32,
     sum: i64,
 }
